replaces old blob log parsing with new module
this parses merge commits properly, and also adds a distinction between src/dest paths in the RawBlobSummary
